Update the Machine Learning category with CS-229
CS-229, offered by the Stanford University is a both
on campus as well as an online learning curriculum.
It takes you through the theory in work behind ML and
steps away from the classical approach of implementing
in abstraction withouth understanding the work that
goes in the back of it. Thus, be prepared for a not-so
hands on grind and hope for a fruitful result!